Allows persistant connections through the proxy
By default the connection will persist on, with a 5 second timeout to
match Apache's timeout value, and much for the same reason, to keep load
down.

It tries to watch out for ErrPersistEOF and tries to handle appropriately,
but I've not witnessed any of these appearing, so I'm not sure it will do.